Navigator changes
