+++ lib/pages/home/widgets/side_nav_bar.dart
@@ -53,6 +53,16 @@ class _SideNavBarState extends State<SideNavBar> {
             onTap: () => widget.onSelectedPage(PageItem.events),
             isActive: widget.currentPage == PageItem.events,
           ),
+          SideNavBarItem(
+            title: 'Map',
+            imagePathActive: 'assets/img/icons/map-filled.png',
+            imagePathInactive: 'assets/img/icons/map-outlined.png',
+            iconHeight: 32,
+            bottomIconPadding: 0,
+            verticalPadding: 7,
+            onTap: () => widget.onSelectedPage(PageItem.pathfinder),
+            isActive: widget.currentPage == PageItem.pathfinder,
+          ),
           // Mensa
           SideNavBarItem(
             title: 'Mensa',
