+++ lib/pages/home/widgets/side_nav_bar_item.dart
@@ -15,6 +15,15 @@ class SideNavBarItem extends StatefulWidget {
   /// ATTENTION: Only use .png-files
   final String imagePathInactive;
 
+  /// Set the icon height
+  final double iconHeight;
+
+  // Set the padding below the icon
+  final double bottomIconPadding;
+
+  // Set the padding at the top and bottom of the whole item
+  final double verticalPadding;
+
   /// Title of the page that this menu item refers to
   final String title;
 
@@ -28,6 +37,9 @@ class SideNavBarItem extends StatefulWidget {
     Key? key,
     required this.imagePathActive,
     required this.imagePathInactive,
+    this.iconHeight = 26,
+    this.bottomIconPadding = 5,
+    this.verticalPadding = 10,
     required this.title,
     required this.onTap,
     this.isActive = false,
@@ -38,8 +50,6 @@ class SideNavBarItem extends StatefulWidget {
 }
 
 class _SideNavBarItemState extends State<SideNavBarItem> {
-  // Adjust this value in order to change the icon height of each navbar-element
-  static const double iconHeight = 26;
   // Adjust this value in order to change the animation curve that is used for the
   // vertical translation-animation
   static const Curve animationCurve = Curves.easeOutExpo;
@@ -56,15 +66,15 @@ class _SideNavBarItemState extends State<SideNavBarItem> {
           CustomButton(
         tapHandler: () => widget.onTap(),
         child: Padding(
-          padding: const EdgeInsets.symmetric(vertical: 10),
+          padding: EdgeInsets.symmetric(vertical: widget.verticalPadding),
           child: Column(
             mainAxisSize: MainAxisSize.min,
             children: [
               Padding(
-                padding: const EdgeInsets.only(left: 14, right: 14, bottom: 5),
+                padding: EdgeInsets.only(left: 14, right: 14, bottom: widget.bottomIconPadding),
                 child: Image.asset(
                   widget.isActive ? widget.imagePathActive : widget.imagePathInactive,
-                  height: iconHeight,
+                  height: widget.iconHeight,
                   color: widget.isActive
                       ? Provider.of<ThemesNotifier>(context).currentThemeData.colorScheme.secondary
                       : Provider.of<ThemesNotifier>(context, listen: false).currentTheme == AppThemes.light
